[OpenCL][Coverity] Add exception handling (#9071)

+++ opencl-intel/backend/libraries/ocl_builtins/utils/TableGen/ClangUtils.cpp
@@ -93,7 +93,7 @@ void build(const std::string &code, std::string fileName) {
 }
 
 // generates 'dummy code' (which does nothing but lets the module compile)
-std::string generateDummyBody(const std::string &type, size_t veclen) {
+std::string generateDummyBody(const std::string &type, unsigned int veclen) {
   std::stringstream sstream;
   sstream << "{return ";
   if ("void" == type) {
